+++ app/controllers/doorkeeper/authorizations_controller.rb
@@ -8,18 +8,12 @@ module Doorkeeper
       if pre_auth.authorizable?
         render_success
       else
-        render_error pre_auth.error_response
+        render_error
       end
-    rescue Errors::DoorkeeperError => e
-      error_response = get_error_response_from_exception(e)
-      render_error error_response
     end
 
     def create
       redirect_or_render authorize_response
-    rescue Errors::DoorkeeperError => e
-      error_response = get_error_response_from_exception(e)
-      redirect_or_render error_response
     end
 
     def destroy
@@ -38,12 +32,12 @@ module Doorkeeper
       end
     end
 
-    def render_error(error_response)
+    def render_error
       if Doorkeeper.configuration.api_only
-        render json: error_response.body,
+        render json: pre_auth.error_response.body,
                status: :bad_request
       else
-        render :error, locals: { error_response: error_response }
+        render :error
       end
     end