+++ spec/lib/oauth/code_request_spec.rb
@@ -13,11 +13,14 @@ module Doorkeeper::OAuth
       client = Doorkeeper::OAuth::Client.new(application)
 
       attributes = {
+        client_id: client.uid,
         response_type: "code",
         redirect_uri: "https://app.com/callback",
       }
 
-      PreAuthorization.new(server, client, attributes)
+      pre_auth = PreAuthorization.new(server, attributes)
+      pre_auth.authorizable?
+      pre_auth
     end
 
     let(:owner) { double :owner, id: 8900 }